reorg files and add stream events feature
