+++ container-manager/src/libs/docker.js
@@ -1,0 +1,28 @@
+// const { execSync } = require('child_process');
+// const fs = require('fs')
+// const path = require('path'); 
+// const axios = require('axios');
+// const mountPath = path.join(__dirname, '../mount/generated/')
+// const instance = axios.create({
+//   socketPath: '/var/run/docker.sock',
+//   baseURL: "http://unix:/", 
+// });
+
+// module.exports = {
+//   initModule,
+//   getSubnetContainers,
+//   testAxios,
+//   getContainersState,
+//   startComposeProfile,
+//   deployContract,
+//   stopComposeProfile,
+// }
+
+// async function testAxios(){
+//   // startSubnetNodes()
+//   const [subnets, services] = await getContainersState()
+//   return {
+//     subnets: subnets,
+//     services: services
+//   }
+// }
