+++ container-manager/src/package.json
@@ -10,6 +10,7 @@
   "dependencies": {
     "axios": "^1.7.9",
     "dockerode": "^3.3.4",
+    "dotenv": "^16.4.7",
     "express": "^4.21.2"
   },
   "devDependencies": {
