+++ container-manager/src/public/script.js
@@ -1,19 +1,54 @@
-function callApi(route) {
-  fetch(route, {
-      method: 'GET', // or 'POST' depending on your API requirements
-      headers: {
-          'Content-Type': 'application/json'
-      }
-  })
-  .then(response => response.json())
-  .then(data => {
-      // Display the formatted JSON response in the output div
-      const outputDiv = document.getElementById('output');
-      outputDiv.textContent = JSON.stringify(data, null, 2); // Pretty-print JSON
-  })
-  .catch((error) => {
-      console.error('Error:', error);
-      const outputDiv = document.getElementById('output');
-      outputDiv.textContent = 'API call failed: ' + error.message;
-  });
+async function callApi(route) {
+  const outputDiv = document.getElementById('output');
+  try {
+    const response = await fetch(route, {method: 'GET'});
+    if (!response.ok) {
+      throw new Error(`HTTP error! Status: ${response.status}`);
+    }
+
+    const contentType = response.headers.get('Content-Type');
+    let data;
+    if (contentType && contentType.includes('application/json')) {
+      console.log('response header JSON')
+      data = await response.json();
+      data = JSON.stringify(data, null, 2);
+    } else {
+      data = await response.text();
+    }
+
+    outputDiv.textContent = data
+  } catch (error){
+    console.error('Error:', error);
+    outputDiv.textContent = 'API call failed: ' + error.message; 
+  }
 }
+
+async function callStreamApi(route){
+  const outputElement = document.getElementById('output'); 
+  const eventSource = new EventSource(route);
+  try {
+    const response = await fetch(route, {method: 'GET'});
+    if (!response.ok) {
+      throw new Error(`HTTP error! Status: ${response.status}`);
+    }
+    outputElement.textContent = ''
+    eventSource.onmessage = (event) => {
+      outputElement.textContent += event.data + '\n'; 
+      outputElement.scrollTop = outputElement.scrollHeight; 
+    };
+
+    eventSource.addEventListener('close', (event) => {
+      outputElement.textContent += event.data + '\n'; 
+      eventSource.close(); 
+    });
+
+    eventSource.onerror = () => {
+      console.log('EventSource failed.');
+      outputElement.textContent += 'Error: Connection lost.\n';
+      eventSource.close();
+    }
+  } catch (error){
+    console.error('Error:', error);
+    outputElement.textContent = 'API call failed: ' + error.message; 
+  }
+};
