+++ container-manager/src/script.sh
@@ -1,0 +1,1 @@
+echo "Hello, World!"
